ref(extractors): Split project and event extractors
